change datatype from floatfield to integerfield

+++ lopa/models.py
@@ -5,7 +5,7 @@ class Cause(models.Model):
     description = models.TextField(blank=True, null=True)       
     initial_frequency = models.FloatField(blank=True, null=True)
     event_id = models.IntegerField(blank=True, null=True)
-    cause_id = models.FloatField(blank=True, null=True)
+    cause_id = models.IntegerField(blank=True, null=True)
     target_frequency = models.FloatField(blank=True, null=True)
 
     class Meta:
@@ -40,7 +40,7 @@ class ConsequenceBarrier(models.Model):
     description = models.TextField(blank=True, null=True)
     pfd = models.FloatField(blank=True, null=True)
     consequence_id = models.IntegerField(blank=True, null=True)
-    consequence_barrier_id = models.FloatField(blank=True, null=True)
+    consequence_barrier_id = models.IntegerField(blank=True, null=True)
 
     class Meta:
         managed = False
@@ -51,7 +51,7 @@ class Event(models.Model):
     description = models.TextField(blank=True, null=True)
     cause_id = models.IntegerField(blank=True, null=True)
     consequence_id = models.IntegerField(blank=True, null=True)
-    event_id = models.FloatField(blank=True, null=True)
+    event_id = models.IntegerField(blank=True, null=True)
 
     class Meta:
         managed = False
